Fix circular dependencies from 'task' extension
- fixed the circular dependencies from the `@theia/task` extension.
