+++ packages/task/src/browser/task-action-provider.ts
@@ -14,54 +14,4 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  ********************************************************************************/
 
-import { injectable, inject } from 'inversify';
-import { TaskService } from './task-service';
-import { TaskRunQuickOpenItem } from './quick-open-task';
-import { QuickOpenBaseAction, QuickOpenItem } from '@theia/core/lib/browser/quick-open';
-import { QuickOpenAction, QuickOpenActionProvider } from '@theia/core/lib/common/quick-open-model';
-import { ThemeService } from '@theia/core/lib/browser/theming';
-
-@injectable()
-export class ConfigureTaskAction extends QuickOpenBaseAction {
-
-    @inject(TaskService)
-    protected readonly taskService: TaskService;
-
-    constructor() {
-        super({ id: 'configure:task' });
-
-        this.updateTheme();
-
-        ThemeService.get().onThemeChange(() => this.updateTheme());
-    }
-
-    async run(item?: QuickOpenItem): Promise<void> {
-        if (item instanceof TaskRunQuickOpenItem) {
-            this.taskService.configure(item.getTask());
-        }
-    }
-
-    protected updateTheme(): void {
-        const theme = ThemeService.get().getCurrentTheme().id;
-        if (theme === 'dark') {
-            this.class = 'quick-open-task-configure-dark';
-        } else if (theme === 'light') {
-            this.class = 'quick-open-task-configure-bright';
-        }
-    }
-}
-
-@injectable()
-export class TaskActionProvider implements QuickOpenActionProvider {
-
-    @inject(ConfigureTaskAction)
-    protected configureTaskAction: ConfigureTaskAction;
-
-    hasActions(): boolean {
-        return true;
-    }
-
-    getActions(): ReadonlyArray<QuickOpenAction> {
-        return [this.configureTaskAction];
-    }
-}
+export { ConfigureTaskAction, TaskActionProvider } from './quick-open-task';
